+++ src/plus/integrations/authentication/gitlab.ts
@@ -1,6 +1,5 @@
 import type { AuthenticationSession, Disposable, QuickInputButton } from 'vscode';
 import { env, ThemeIcon, Uri, window } from 'vscode';
-import { supportedInVSCodeVersion } from '../../../system/utils';
 import type {
 	IntegrationAuthenticationProvider,
 	IntegrationAuthenticationSessionDescriptor,
@@ -53,11 +52,9 @@ export class GitLabAuthenticationProvider implements IntegrationAuthenticationPr
 				input.password = true;
 				input.title = `GitLab Authentication${descriptor?.domain ? `  \u2022 ${descriptor.domain}` : ''}`;
 				input.placeholder = `Requires ${descriptor?.scopes.join(', ') ?? 'all'} scopes`;
-				input.prompt = supportedInVSCodeVersion('input-prompt-links')
-					? `Paste your [GitLab Personal Access Token](https://${
-							descriptor?.domain ?? 'gitlab.com'
-					  }/-/profile/personal_access_tokens "Get your GitLab Access Token")`
-					: 'Paste your GitLab Personal Access Token';
+				input.prompt = `Paste your [GitLab Personal Access Token](https://${
+					descriptor?.domain ?? 'gitlab.com'
+				}/-/profile/personal_access_tokens "Get your GitLab Access Token")`;
 				input.buttons = [infoButton];
 
 				input.show();